requirements.txt
asgiref==3.8.1
Command==0.1.0
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.2
psycopg==3.1.18
setuptools==70.2.0
sqlparse==0.5.0
typing_extensions==4.11.0
wheel==0.43.0
